requirements.txt
asgiref==3.5.2
attrs==21.4.0
Automat==20.2.0
black==22.3.0
certifi==2022.6.15
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
constantly==15.1.0
cryptography==37.0.2
cssselect==1.1.0
dj-database-url==0.5.0
Django==4.0.5
django-cors-headers==3.13.0
django-crispy-forms==1.14.0
django-filter==22.1
django-heroku==0.3.1
djangorestframework==3.13.1
filelock==3.7.1
gunicorn==20.1.0
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
itemadapter==0.6.0
itemloaders==1.0.4
jmespath==1.0.0
lxml==4.9.0
Markdown==3.3.7
mypy-extensions==0.4.3
parsel==1.6.0
pathspec==0.9.0
platformdirs==2.5.2
Protego==0.2.1
psycopg2==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyDispatcher==2.0.5
pyOpenSSL==22.0.0
python-dotenv==0.20.0
python-scrapyd-api==2.1.2
pytz==2022.1
queuelib==1.6.2
requests==2.28.0
requests-file==1.5.1
Scrapy==2.6.1
scrapy-djangoitem==1.1.1
scrapyd==1.3.0
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.2
tldextract==3.3.0
tomli==2.0.1
Twisted==22.4.0
typing_extensions==4.2.0
urllib3==1.26.9
w3lib==1.22.0
whitenoise==6.2.0
zope.interface==5.4.0
